deploy with mock added
