+++ scripts/actions/getRoleSingers.js
