+++ scripts/deploy-mock.js
@@ -1,150 +1,9 @@
 const hre = require("hardhat");
-const ethers = hre.ethers;
-const fs = require("fs");
-const path = require("path");
-const deployContracts = require("./actions/deployContracts");
-const getRoleSigners = require("./utils/getRoleSingers");
-const { address, hashIt } = require("./utils/helpers");
-const grantRoles = require("./utils/grantRoles");
-const storeConfigs = require("./utils/storeConfigs");
 
-const pipeActions = require("./utils/pipeActions");
-
-const initializer = require("./actions/initializer");
-const getContractFactories = require("./actions/getContractFactories");
-
-const setupAndDeploy = require("./actions/setupAndDeploy");
-const mockMint = require("./actions/mockMint");
+const deployWithMock = require("./actions/deployWithMock");
 
 async function main() {
-  const signers = await ethers.getSigners();
-
-  /// === Contract Factories
-
-  /*const {
-    GlobalAccessControl,
-    CitadelToken,
-    StakedCitadelVester,
-    StakedCitadel,
-    StakedCitadelLocker,
-    SupplySchedule,
-    CitadelMinter,
-    KnightingRound,
-    Funding,
-    wBTC,
-    CVX,
-    USDC,
-  } = await getContractFactories();*/
-
-  /// === Deploying Contracts & loggin addresses
-  /* const {
-    gac,
-    citadel,
-    xCitadel,
-    xCitadelVester,
-    xCitadelLocker,
-    schedule,
-    citadelMinter,
-    knightingRound,
-    fundingWbtc,
-    fundingCvx,
-    wbtc,
-    cvx,
-    usdc,
-  } = await deployContracts([
-    { factory: GlobalAccessControl, instance: "gac" },
-    { factory: CitadelToken, instance: "citadel" },
-    { factory: StakedCitadel, instance: "xCitadel" },
-    { factory: StakedCitadelVester, instance: "xCitadelVester" },
-    { factory: StakedCitadelLocker, instance: "xCitadelLocker" },
-    { factory: SupplySchedule, instance: "schedule" },
-    { factory: CitadelMinter, instance: "citadelMinter" },
-    { factory: KnightingRound, instance: "knightingRound" },
-    { factory: Funding, instance: "fundingWbtc" },
-    { factory: Funding, instance: "fundingCvx" },
-    { factory: wBTC, instance: "wbtc" },
-    { factory: CVX, instance: "cvx" },
-    { factory: USDC, instance: "usdc" },
-  ]);*/
-
-  /// === Variable Setup
-
-  /*  const {
-    governance,
-    keeper,
-    guardian,
-    treasuryVault,
-    techOps,
-    treasuryOps,
-    citadelTree,
-    policyOps,
-    eoaOracle,
-  } = await getRoleSigners(); */
-
-  const mintTo = signers[0].address;
-  const xCitadelFees = [0, 0, 0, 0];
-
-  await pipeActions({ xCitadelFees, mintTo })(
-    setupAndDeploy,
-    () => console.log("Contracts setted up ..."),
-    mockMint,
-    () => console.log("Mock mints ready ..."),
-    getRoleSigners,
-    () => console.log("Roles assigned ..."),
-    initializer,
-    () => console.log("Contracts initialized ...")
-  );
-
-  /// === Initialization and Setup
-
-  /* const xCitadelFees = [0, 0, 0, 0];
-
-  await initializer({
-    gac,
-    citadel,
-    xCitadel,
-    xCitadelVester,
-    xCitadelLocker,
-    citadelMinter,
-    schedule,
-    fundingWbtc,
-    fundingCvx,
-  })({
-    governance,
-    xCitadelFees,
-    keeper,
-    guardian,
-    treasuryVault,
-    techOps,
-    citadelTree,
-    wbtc,
-    cvx,
-    eoaOracle,
-  });
-
-  /// ======== Grant roles
-
-  await grantRoles(gac, governance, getRoleSigners, { citadelMinter });
-
-  /// =================================== ///
-  // Storing the contract addresses for accessing in helper scripts
-
-  storeConfigs(
-    path.join(__dirname, "..", "scripts-data"),
-    "testnet-addresses",
-    {
-      gac: address(gac),
-      citadel: address(citadel),
-      xCitadel: address(xCitadel),
-      xCitadelVester: address(xCitadelVester),
-      xCitadelLocker: address(xCitadelLocker),
-      schedule: address(schedule),
-      citadelMinter: address(citadelMinter),
-      knightingRound: address(knightingRound),
-      wbtc: address(wbtc),
-      cvx: address(cvx),
-    }
-  );*/
+  await deployWithMock();
 }
 
 main()
